backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
requests==2.31.0
django-cors-headers==4.3.1
flower==2.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0